+++ source/Genome.h
@@ -45,8 +45,10 @@ class Genome {
         uint genomeInsertL; //total length of the sequence to be inserted on the fly
         uint genomeInsertChrIndFirst; //index of the first inserted chromosome
 
+        SharedMemory * sharedMemory;
+        
         Genome (Parameters &Pin );
-        ~Genome();
+        //~Genome();
 
         void freeMemory();
         void genomeLoad();
@@ -61,8 +63,7 @@ class Genome {
         Parameters &P;
         key_t shmKey;
         char *shmStart;
-        char *G1; //pointer -200 of G
-        SharedMemory * sharedMemory;
+        char *G1; //pointer -200 of G        
         uint OpenStream(string name, ifstream & stream, uint size);
         void HandleSharedMemoryException(const SharedMemoryException & exc, uint64 shmSize);