+++ source/SequenceFuns.cpp
@@ -127,18 +127,47 @@ void nuclPackBAM(char* ReadsIn, char* ReadsOut, uint Lread) {//pack nucleotides 
     };
 };
 
-void convertNucleotidesToNumbers(const char* R0, char* R1, uint Lread) {//transform sequence  from ACGT into 0-1-2-3 code
+void convertNucleotidesToNumbers(const char* R0, char* R1, const uint Lread) {//transform sequence  from ACGT into 0-1-2-3 code
     for (uint jj=0;jj<Lread;jj++) {
                     switch (int(R0[jj])){
-                        case (65): case(97):  R1[jj]=char(0);break;//A
-                        case (67): case(99):  R1[jj]=char(1);break;//C
-                        case (71): case(103): R1[jj]=char(2);break;//G
-                        case (84): case(116): R1[jj]=char(3);break;//T
-//                         case (78): R1[jj]=char(9);break;//N
-                        default:   R1[jj]=char(9);//anything else
+                        case (65): case(97):  
+                            R1[jj]=char(0);break;//A
+                        case (67): case(99):  
+                            R1[jj]=char(1);break;//C
+                        case (71): case(103): 
+                            R1[jj]=char(2);break;//G
+                        case (84): case(116): 
+                            R1[jj]=char(3);break;//T
+                        default:                              
+                            R1[jj]=char(9);//anything else
                     };
                 };
 };
+
+uint convertNucleotidesToNumbersRemoveControls(const char* R0, char* R1, const uint Lread) {//transform sequence  from ACGT into 0-1-2-3 code
+    uint iR1=0;
+    for (uint jj=0;jj<Lread;jj++) {
+        switch (int(R0[jj])){
+            case (65): case(97):  
+                R1[jj]=char(0);break;//A
+            case (67): case(99):  
+                R1[jj]=char(1);break;//C
+            case (71): case(103): 
+                R1[jj]=char(2);break;//G
+            case (84): case(116): 
+                R1[jj]=char(3);break;//T
+            default:                              
+                if (int(R0[jj]) < 32) {//control characters are skipped
+                    continue;
+                } else {//all non-control non-ACGT characters are convreted to N
+                    R1[jj]=char(9);//anything else
+                };
+        };
+        ++iR1;
+    };
+    return iR1;
+};
+
 
 char convertNt01234(const char R0) {//transform sequence  from ACGT into 0-1-2-3 code    
     switch(R0)